Fix UserDefaults settings source-less observation
